optim (JokeService): is now requesting once
Previously JokeService was requesting the joke.json file each time user
change page (going from 'today' to 'list' dor example). This was a
problem because the file do not change often so it is not needed to
request it even if web browsers are using cache system. I saw that file
was reloaded whe URL changes.